Move backend code to /server directory
