+++ Makefile
@@ -2,8 +2,8 @@
 
 build: gomodgen
 	export GO111MODULE=on
-	env GOARCH=amd64 GOOS=linux go build -ldflags="-s -w" -o bin/findMovie handlers/findMovie/main.go
-	env GOARCH=amd64 GOOS=linux go build -ldflags="-s -w" -o bin/findShow handlers/findShow/main.go
+	env GOARCH=amd64 GOOS=linux go build -ldflags="-s -w" -o bin/findMovie server/handlers/findMovie/main.go
+	env GOARCH=amd64 GOOS=linux go build -ldflags="-s -w" -o bin/findShow server/handlers/findShow/main.go
 
 clean:
 	rm -rf ./bin ./vendor go.sum
