+++ server/db/mongo/instance.go
@@ -3,7 +3,7 @@ package mongo
 import (
 	"context"
 	"fmt"
-	"github.com/braydend/movie-list/utils"
+	"github.com/braydend/movie-list/server/utils"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"log"
